requirements.txt
beautifulsoup4==4.5.1
blessings==1.6
bpython==0.15
curtsies==0.2.6
greenlet==0.4.10
GroupyAPI==0.8.1
inflect==0.2.5
inflection==0.3.1
python-dotenv==0.8.2
requests==2.9.1
six==1.10.0
wcwidth==0.1.7
wikipedia==1.4.0
